Backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
django-filter==25.1
Pillow==11.3.0
psycopg2-binary==2.9.10 